CHanges requested by PO

+++ src/Controller/AdminProductController.php
@@ -119,12 +119,19 @@ class AdminProductController extends AbstractController
         return $this->twig->render('Admin/add.html.twig', ['categories' => $categories, 'post' => true]);
     }
 
-    public function delete(int $id)
+    /**
+     *
+     * Delete product from bdd
+     */
+    public function delete()
     {
-        $pictureManager = new PictureManager();
-        $pictureManager->delete($id);
-        $productManager = new ProductManager();
-        $productManager->delete($id);
-        header('location: /adminProduct/list');
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            $id = $_POST['id'];
+            $pictureManager = new PictureManager();
+            $pictureManager->delete($id);
+            $productManager = new ProductManager();
+            $productManager->delete($id);
+            header('location: /adminProduct/list');
+        }
     }
 }
